requirements.txt
amqp==5.3.1
arrow==1.3.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==23.1.2
Automat==24.8.1
backports.zoneinfo==0.2.1
billiard==4.2.1
binaryornot==0.4.4
boto3==1.37.16
botocore==1.37.16
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels-redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cookiecutter==2.6.0
cryptography==44.0.2
daphne==4.1.2
Django==4.2.20
django-allauth==65.6.0
django-browser-reload==1.15.0
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-environ==0.11.2
django-formtools==2.5.1
django-mpesa==2.0.14
django-otp==1.5.4
django-phonenumber-field==8.0.0
django-tailwind==3.6.0
django-two-factor-auth==1.17.0
djangorestframework==3.15.2
dotenv==0.9.9
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
intasend-python==1.1.2
jinja2==3.1.6
jmespath==1.0.1
jwt==1.3.1
kombu==5.5.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.1.0
paypalrestsdk==1.13.3
phonenumbers==9.0.2
pillow==10.4.0
prompt-toolkit==3.0.50
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pygments==2.19.1
pyOpenSSL==25.0.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
qrcode==7.4.2
razorpay==1.4.2
redis==5.2.1
requests==2.32.3
rich==13.9.4
s3transfer==0.11.4
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
stripe==11.6.0
text-unidecode==1.3
tomli==2.2.1
twisted==24.11.0
txaio==23.1.1
types-python-dateutil==2.9.0.20241206
typing-extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
zope.interface==7.2
